requirements.txt
aioredis==1.3.1
asgiref==3.2.10
async-timeout==3.0.1
attrs==19.3.0
autobahn==20.6.1
Automat==20.2.0
CacheControl==0.12.6
cachetools==4.1.0
certifi==2020.4.5.2
cffi==1.14.0
chardet==3.0.4
constantly==15.1.0
cryptography==2.9.2
daphne==2.5.0
Django==3.1.1
django-cors-headers==3.3.0
django-filebrowser==3.13.1
django-gcloud-storage==0.4.0
django-grappelli==2.14.2
django-rest-knox==4.1.0
django-storages==1.9.1
djangorestframework==3.11.0
docx==0.2.4
drf-firebase-auth==0.0.8
et-xmlfile==1.0.1
Faker==4.1.2
firebase-admin==4.3.0
google-api-core==1.20.0
google-api-python-client==1.10.1
google-auth==1.17.2
google-auth-httplib2==0.0.4
google-cloud-core==1.4.1
google-cloud-firestore==1.9.0
google-cloud-storage==1.29.0
google-cloud-vision==1.0.0
google-resumable-media==0.5.1
googleapis-common-protos==1.52.0
grpcio==1.31.0
hiredis==1.0.1
httplib2==0.18.1
hyperlink==19.0.0
idna==2.9
incremental==17.5.0
jdcal==1.4.1
lxml==4.5.1
msgpack==0.6.2
numpy==1.18.5
openpyxl==3.0.3
pandas==1.0.4
Pillow==7.1.2
protobuf==3.12.2
psycopg2==2.8.6
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
PyHamcrest==2.0.2
pyOpenSSL==19.1.0
python-dateutil==2.8.1
python-docx==0.8.10
pytz==2020.1
requests==2.23.0
rsa==4.6
service-identity==18.1.0
six==1.15.0
sqlparse==0.3.1
text-unidecode==1.3
Twisted==20.3.0
txaio==20.4.1
uritemplate==3.0.1
urllib3==1.25.9
xlrd==2.0.1
zope.interface==5.1.0
